Improve no_std doc again, mostly clarity
